fix(#202): Fix extension artifact upload via Snap
- Update snap version to v0.0.8
